refactor(scheduling): extract useWeekStaffingSuggestions to shared hook
Move the local useWeekStaffingSuggestions function from StaffingOverlay.tsx
into src/hooks/useWeekStaffingSuggestions.ts so the timeline view (Task 9)
can call it directly. Exports WeekStaffingSuggestions (ReturnType) for
downstream consumers. No behavior change — all 13 StaffingOverlay tests pass.
